+++ crates/tests/src/deviations.rs
@@ -1,6 +1,7 @@
 //! Known deviations (or not) from the Definition.
 
 mod implicit_ty_var;
+mod lunar_ml;
 mod mlton;
 mod smlnj;
 mod successor_ml;